config: Add GITMIRROR option.
This is so that we can build even when remote git server
is down.

It is possible to use other git mirror than git@public.inteno.se
but the sanity checks will fail and it won't be possible
to mirror the inteno git repos as the mirror is disabled
for inteno servers.

sdk: Ability to build without an Inteno Git mirror.

Fix git_clone.sh for public build

Public build fails to clone from gitmirror because rewrite is
not set.

config: always set inteno mirror

scripts: fix basename filter if URL argument was a SSH reference instead of HTTP.

scripts: fix rewrite logic