Reorganize BCM config options [iopsyswrt]
Unused options have been removed.
Build options for the bcmkernel package have been moved to the bcmkernel
KConfig.
Options pertaining the image build have been moved to the image submenu
of the target KConfig.

Sometimes the lines are a bit blurry what constitutes a target option
and what constitutes a bcmkernel option, because bcmkernel is not just
an ordinary package and also deals with other things such as parts of
the image generation, which is more of less an implementation detail. In
general, image related option are are TARGET_ options while build
related options are left to BCM_.

Renamed/moved
* TARGET_NO_VOICE => !BCM_VOICE
* TARGET_NO_DSL   => !BCM_DSL
* TARGET_DSL_*    => BCM_DSL_*

Removed, not used
* BCM_BLNAND
* BCM_BLUETOOTH
* BCM_CFE_PASSWORD
* BCM_I2C
* BCM_USBSOUND_MODULES
* SECURE_BOOT_CFE
* TARGET_CFERAM_EXT
* TARGET_NO_BLUETOOTH
* TARGET_NO_DECT (none of our targets has internal DECT)
* TARGET_NO_DHD
* TARGET_CFERAM_EXT