trusted-firmware-a.mk: Remove BUILD_STRING var
BUILD_STRING includes spaces that are are escaped when calling
following make with BUILD_STRING as a parameter.

After this operation BUILD_STRING looks like:
"OpenWrt\ ve43514ab-4\ (mt7986-snand)"
This parameter is then used as a part of a VERSION_STRING
in compilation of build_message.o (make_helpers/build_macros.mk:522).
This generates an error:
<stdin>:1:144: error: unknown escape sequence: '\040' [-Werror]

If BUILD_STRING is not defined it will be generated from
branch/commit ID by ATF makefile.