openssl: set ARM CPU capabilities at compile time
BCM63138, BCM6846 and BCM6878 are asymmetric and have VFP/NEON support
on CPU 0 only. The NEON support on CPU 0 is intended for use with VoIP
and cannot be used for anything else as it is unavailable on CPU 1 and
in addition VFP/NEON registers are not saved/restored on context
switch. It is therefore necessary to disable auto-detection of NEON
support in openssl for these SoCs.