Revert "scripts: fix for curl timeout"
This reverts commit 3c52aa25a1ad7d34c3d867eaafb7caba61842b5f.

The reverted commit made FTP downloads more efficient using the curl
option `--ftp-method singlecwd`.

https://curl.se/docs/manpage.html#--ftp-method

According to the commit message, it solved some timeout issue, but it
unclear which specific package download the fix was for.  Given that it
is not clear whether the patch is needed at all, FTP is only used very
sparingly these days and HTTP mirrors are available, it is better to
revert it rather than maintaining this as a downstream patch when it is
most likely no longer needed.

Signed-off-by: Andreas Gnau <andreas.gnau@iopsys.eu>